Move global variables to globalvars
