+++ src/piidigger/__main__.py
@@ -3,7 +3,7 @@ import traceback
 from multiprocessing import Process, freeze_support, set_start_method
 
 from piidigger.piidigger import main
-from piidigger.globalfuncs import errorCodes
+from piidigger.globalvars import errorCodes
 
 exitCode = errorCodes['ok']
 
@@ -17,7 +17,7 @@ if __name__=='__main__':
     except KeyboardInterrupt:
         pass
     except Exception:
-        exit_code = errorCodes['unknownError']
+        exitCode = errorCodes['unknownError']
         errorFile='piidigger.exc'
         print(f'An unknown error was encountered.  Detailed error information has been written to {errorFile}.')
         traceback.print_exception(file=errorFile)
