+++ src/piidigger/globalfuncs.py
@@ -10,20 +10,8 @@ from piidigger import console
 from piidigger import filehandlers as fh
 from piidigger import datahandlers as dh
 from piidigger import outputhandlers as oh
+from piidigger.globalvars import maxChunkSize
 
-######################################################
-############### Global variables #####################
-######################################################
-
-errorCodes={
-    "ok": 0,
-    "userTerm": 1,
-    "invalidConfig": 1,
-    'unknownError': 1024,
-}
-
-excelBlankRowLimit=250
-excelBlankColLimit=500
 
 # Dynamically build the supported file handlers based on the contents of the filehandlers package.
 # Each file handler needs a globally-defined variable called "handles" with a dictionary as follows:
@@ -251,7 +239,7 @@ def isAdmin() -> bool:
     return check
 
 
-def makeChunks(s: str, chunkSize: int=650) -> list:
+def makeChunks(s: str, chunkSize: int=maxChunkSize) -> list:
     '''Breaks up a string into smaller strings not larger than chunkSize'''
 
     words = s.split()
