+++ src/piidigger/globalvars.py
@@ -1,0 +1,11 @@
+errorCodes={
+    "ok": 0,
+    "userTerm": 1,
+    "invalidConfig": 1,
+    'unknownError': 1024,
+}
+
+
+excelBlankRowLimit=250
+excelBlankColLimit=500
+maxChunkSize = 650
